debos: rootfs: fix password reset requirement
Existing code arranges to set up an initial well-known password but with
a requirement to reset it on first login.

This wasn't working because the well-known password was being set after
expiring the "old" password, meaning that it was being treated as having
been reset immediately.

Instead, ensure to expire the password _after_ setting it.

Fixes: #69
Signed-off-by: Robie Basak <robibasa@qti.qualcomm.com>