ci: add qemu test mechanism
Add a mechanism to test behaviours don't relate to hardware. These can
be effectively tested without need of the target platform using qemu.

This uses Python and pytest. Specific tests to follow.

For now, this is just the raw test that will need to be run manually,
and depends on the developer already having built a disk-ufs.img by
hand using the instructions in README.md. Automation and integration
into GitHub Actions will follow. How-to documentation would be useful
but would be superseded by that automation and integration, so this is
omitted for now.

Signed-off-by: Robie Basak <robibasa@qti.qualcomm.com>