workflows: scan all shell scripts
Instead of just scripts/*.sh, scan all scripts found in the repository
using `file`.

I didn't find a widely accepted mechanism to do this so settled on using
`file`. The script is at least not too long to follow, doesn't introduce
any external dependency, and is careful to avoid errors from unusual
characters in filenames (eg. backslash escapes and carriage returns) to
try and be safe against all inputs.

If this pattern works, we could extend it to the other lint checks in
this file.

Signed-off-by: Robie Basak <robibasa@qti.qualcomm.com>