Add local tests to build system and CI
Add a "make test" target to the Makefile and to CI. This is intended to
run all tests that can be run locally, with just one for now.

For now, knowledge of the specific artifacts required is duplicated
between the Makefile and .github/workflows/test.yml. I don't like this,
but this is currently necessary because the workflows are split and save
and restore artifacts between building and testing. Perhaps this can be
improved, but I'm deferring this for a possible future overhaul.

Signed-off-by: Robie Basak <robibasa@qti.qualcomm.com>