Add local tests to build system and CI
Add a "make test" target to the Makefile and to CI. This is intended to
run all tests that can be run locally, with just one for now.

These tests run quickly, locally and independently of LAVA, so are
bundled in with the debos recipe runs themselves in debos.yml.

Signed-off-by: Robie Basak <robibasa@qti.qualcomm.com>