CI: scan all Python scripts
Now that commit b1781ef established a pattern to find all shell scripts
for linting, do the same for all Python scripts, refactoring the pattern
out into a reusable shell function.

This way, adding a file anywhere in the repository will be sufficient to
ensure it is scanned, instead of hoping that we remember to amend the
lint invocations.

Signed-off-by: Robie Basak <robibasa@qti.qualcomm.com>