Verify sha256sums of some binary dependencies
debos has an outstanding feature request to make this integrated in
go-debos/debos#120. In the meantime, we can do this manually by invoking
sha256sum in check mode directly.

We're downloading qcom-ptool directly from the tip of the main branch,
so verifying that download seems less practical. In future, we could tag
a release there, and then pin to that, at which point verifying the hash
of the result would make more sense.

Signed-off-by: Robie Basak <robibasa@qti.qualcomm.com>