ci/schemacheck.py: make pylint clean
We're already using pylint in CI against this repository, but it
currently does not cover this file by accident. To fix that, ideally
this file would comply with our CI policy on everything being pylint
clean.

However, it's difficult currently with our GitHub workflow arrangements
to ensure that all imports for all scripts are available at the time
pylint runs. Since there are only two instances of this in the entire
repository, just skip these two specific import checks for now.

I would like to move these checks so they run locally as well with "make
check", at which point we'd switch the GitHub workflows to doing that
instead, and then everything would be aligned including a single place
to define development dependencies. However, these dependencies are not
currently available as packages, so it seems too much to expect
developers to manage at the moment. That's a rabbithole, so for now, to
make progress, skipping these for the sake of two ignore directives
seems like a reasonable compromise.

Signed-off-by: Robie Basak <robie.basak@oss.qualcomm.com>