qemu_test: log stdout on failure
If the test times out or fails in CI, then it would be useful to see
everything that was output so that we can either fix a bug or adjust our
expect script. Asking pexpect to write to stdout works well for this as
pytest captures test output and only prints it on individual test
failure.

Signed-off-by: Robie Basak <robie.basak@oss.qualcomm.com>