docs(README): Use shields.io to fix broken badges
GitHub badges require authentication for projects that one is a member
of. Members of the project who should really see these badges will have
to login with GitHub before seeing these. Use shields.io as a badge
proxying and caching service.

Use this opportunity to modernize the badge and list the mainline
builds.

Signed-off-by: Loïc Minier <loic.minier@oss.qualcomm.com>