qemu_test: factor out vm fixture
In a future change I want to pass around another value as well as the
pexpect.spawn object, so make the fixture return a SimpleNamespace that
has the spawn object under that instead of it being the direct value.
Change the existing test accordingly.

This should result in no functional change to the test suite.

Signed-off-by: Robie Basak <robie.basak@oss.qualcomm.com>