qemu_test: re-use the same qemu vm fixture
Since qemu booting is slow and we want fast developer iteration, we make
the optimisation compromise that we will not reset the qemu vm test
fixture from a fresh image for every test. Most tests should not collide
with each other. If we think a new test will do that and we want to make
the compromise of giving it an isolated environment for a slower test
suite, we can deal with that then.

Signed-off-by: Robie Basak <robie.basak@oss.qualcomm.com>