qemu_guest_test: add test for sudo fqdn issue
If DNS is misconfigured in the deployment environment, sudo may hang
while it times out on DNS requests, which is not desirable for the
majority of our users. Add a test that identifies if sudo is using DNS.

See: https://github.com/qualcomm-linux/qcom-deb-images/issues/193

There isn't an easy and obvious method of testing "did sudo use DNS"
that (ideally) is independent of the DNS configuration in the test
environment. This is hopefully unintrusive and reasonably flexible.

The catch is that if we use sudo-rs in the future then it might stop
working if Rust doesn't use libc. But this seems better than nothing.

To make it easier to ensure that this test is correctly testing what we
need, I've separated the test from the fix using xfail. This way, before
the issue is fixed, the test suite passing means that the test is
correctly detecting the issue. I'll then introduce the fix in the next
change.

Signed-off-by: Robie Basak <robie.basak@oss.qualcomm.com>