fix(debos/flash): Add dtb for RB1
In the current boot firmware for RB1 images, a DTB is stored in the
u-boot binary which then loads the latest version from the ESP. However
qcom-ptool was modified to align the partition scheme in anticipation of
an upcoming switch to a regular EDK2 based firmware. Generate the
dtb.bin with the correct dtb in anticipation of newer ptool and upcoming
firmware.

Signed-off-by: Loïc Minier <loic.minier@oss.qualcomm.com>