ci: auto detect kernel package
Add support to request the automatic detection and use of the kernel
matching the pattern linux-image-* that is present in the local
directory of extra debs, and use this from the linux.yml workflow.

This saves us having to manually keep linux.yml in sync with the
latest upstream tag.

Discussion in #216

Signed-off-by: Robie Basak <robie.basak@oss.qualcomm.com>